Create a simple bitmap font reader and renderer
There isn't much reason for the editor to run in a terminal window, so
the next step is to draw and get input events from a graphical window
using SDL.

This commit adds a basic test program that can read a font in the BDF
format and display text to the screen. This code will eventually be
integrated with the rest of the editor.