Get editor size from the window instead of the terminal
Previously the editor would resize itself upon startup by finding the
size of the current terminal window, but now that it is running in an
actual window it should get the number of rows and columns from there.

This also adds parameters for the window size in rows and columns to the
window_init() function.