Improve behaviour when opening a file from the command line
The editor had two issues when reading a file with a name from the
command line. The most annoying issue was that when reallocating the
internal filename in the editor, the null terminator wasn't copied which
sometimes left garbage at the end of the filename.

The other issue was that the editor would quit with a fatal error if
there was no file with the specified name, when it should open the file
anyway and just create it on save.