Refactor text rendering code
Previously, the editor would print visible lines plus the statusline and
message bar into a buffer, then the window would display this entire
buffer. This refactor makes the window directly read lines from the
editor and individually display them on the screen. This will make it
easier to do syntax highlighting, since that information is stored per
line. The statusline and message line are still printed into a buffer to
be displayed by the window, but they are now positioned from the bottom
of the screen.